students-project-loc-hong/requirements.txt
langchain==0.1.16
langchain-openai==0.1.4
streamlit==1.33.0
python-dotenv==1.0.1
loguru==0.7.2
watchdog==4.0.0
python-docx==1.1.2